+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Pre-release]
 - Add github actions to automate tag and release
+- Replace DS Header component with dgx-header-component
 
 ## [0.15.1]