+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # CHANGE LOG
 
+## [0.16.0]
+- Add github actions to automate tag and release
+- Replace DS Header component with dgx-header-component v2.8.0-r16-rc-2
+
 ## [0.15.1]
+
 - Fix absolute link is prefixed by host issue
 
 ## [0.15.0]
 
+- add features and tests for advanced search
 - add features and tests for item details pages
 - add features and tests for EDD requests
 - add features and tests for item details pages